% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl.R
\name{bl}
\alias{bl}
\alias{check_bl}
\title{An S4 class to represent a Bayes linear object.}
\usage{
check_bl(object)
}
\description{
An S4 class to represent a Bayes linear object.
}
\section{Slots}{

\describe{
\item{\code{name}}{A string for the name of the variable collection.}

\item{\code{varnames}}{A character vector of variable names.}

\item{\code{expectation}}{A numeric vector of expectations.}

\item{\code{covariance}}{A numeric matrix or a numeric scalar for covariances.}
}}

