% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_adjust.R
\name{bl_adjust}
\alias{bl_adjust}
\title{Bayes linear kinematic adjustment}
\usage{
bl_adjust(x, y)
}
\arguments{
\item{x}{bl object to be adjusted.}

\item{y}{bl or bl_data object to adjust by.}
}
\value{
Adjusted bl object.
}
\description{
Bayes linear kinematic adjustment
}
\examples{
bl1 <- bl(name = 'Variables',
        varnames = c('x', 'y', 'z'),
        expectation = c(1, 2, 3),
        covariance = matrix(c(1, 0.5, 0.5,
        0.5, 1, 0.5,
        0.5, 0.5, 1), 3, 3))
        
bl2 <- bl_data(name = 'Data',
           varnames = c('x', 'y'),
           values = c(1.1, 2.1))

bl_adjust(bl1, bl2)
}
