% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_data.R
\docType{class}
\name{bl_data}
\alias{bl_data}
\title{An S4 class to represent data for a Bayes linear adjustment.}
\description{
An S4 class to represent data for a Bayes linear adjustment.
}
\section{Slots}{

\describe{
\item{\code{name}}{A string for the name of the variable collection.}

\item{\code{varnames}}{A character vector of variable names.}

\item{\code{values}}{A numeric vector of observed values.}
}}

