% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_resolution.R
\name{bl_resolution}
\alias{bl_resolution}
\title{Bayes linear kinematic adjustment}
\usage{
bl_resolution(x, y)
}
\arguments{
\item{x}{bl object prior to adjustment.}

\item{y}{bl object after adjustment.}
}
\value{
Vector of resolutions on the \link{0,1} scale.
}
\description{
Bayes linear kinematic adjustment
}
