% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_subset.R
\name{bl_subset}
\alias{bl_subset}
\title{bl object subsetting}
\usage{
bl_subset(x, varnames)
}
\arguments{
\item{x}{bl object to be subsetted.}

\item{varnames}{variable names to be extracted.}
}
\value{
Extracted bl object.
}
\description{
bl object subsetting
}
