% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellinger_squared.R
\name{hellinger_squared}
\alias{hellinger_squared}
\title{Squared Hellinger distance between two bl objects}
\usage{
hellinger_squared(x, y)
}
\arguments{
\item{x}{bl object.}

\item{y}{bl object.}
}
\value{
Hellinger distance on the \link{0,1} scale.
}
\description{
Squared Hellinger distance between two bl objects
}
