% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl.R
\name{plot,bl-method}
\alias{plot,bl-method}
\title{Plot method for bl class ('plot')}
\usage{
\S4method{plot}{bl}(x)
}
\arguments{
\item{x}{The object of class 'bl' to plot.}
}
\description{
Plots the covariance matrix stored in a 'bl' object. If the covariance
matrix is a scalar, a message is printed to the console.
}
