% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_data.R
\name{print,bl_data-method}
\alias{print,bl_data-method}
\title{Print method for bl_data class}
\usage{
\S4method{print}{bl_data}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{The object of class 'bl_data' to print.}

\item{digits}{Minimum number of significant digits to display (default:
\code{getOption("digits")}). Passed to \code{round()}.}

\item{...}{Further arguments passed to or from other methods (currently unused).}
}
\value{
Invisibly returns the original object \code{x}.
}
\description{
Displays the observed values stored in a Bayes linear data ('bl_data') object,
allowing control over the number of digits shown. Called automatically when
an object of class 'bl_data' is printed to the console.
}
\examples{
data1 <- bl_data(name = 'Experiment A',
                 varnames = c('Temp', 'Pressure'),
                 values = c(25.123, 1013.456))
data1
print(data1, digits = 4)

}
