% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl.R
\name{show,bl-method}
\alias{show,bl-method}
\title{Basic show method for bl class}
\usage{
\S4method{show}{bl}(object)
}
\arguments{
\item{object}{The object of class 'bl' to show.}
}
\value{
Invisibly returns the original object \code{object}.
}
\description{
Displays the contents of a Bayes linear ('bl') object. Called automatically
when an object of class 'bl' is shown in the console.
}
