% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_data.R
\name{show,bl_data-method}
\alias{show,bl_data-method}
\title{Basic show method for bl_data class}
\usage{
\S4method{show}{bl_data}(object)
}
\arguments{
\item{object}{The object of class 'bl_data' to show.}
}
\value{
Invisibly returns the original object \code{object}.
}
\description{
Displays the contents of a Bayes linear data ('bl_data') object.
}
