% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl.R
\docType{class}
\name{bl}
\alias{bl}
\title{An S4 class to represent a Bayes linear object.}
\description{
An S4 class to represent a Bayes linear object.
}
\section{Slots}{

\describe{
\item{\code{name}}{A string for the name of the variable collection.}

\item{\code{varnames}}{A character vector of variable names.}

\item{\code{expectation}}{A numeric vector of expectations.}

\item{\code{covariance}}{A numeric matrix or a numeric scalar for covariances.}
}}

\examples{
bl1 <- bl(name = 'Example 1',
          varnames = c('x', 'y', 'z'),
          expectation = c(1, 2, 3),
          covariance = matrix(c(1, 0.5, 0.5,
                      0.5, 1, 0.5,
                      0.5, 0.5, 1), 3, 3))
                      
}
