% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_adjust.R
\name{bl_adjust}
\alias{bl_adjust}
\title{Bayes linear kinematic adjustment}
\usage{
bl_adjust(x, y)
}
\arguments{
\item{x}{bl object to be adjusted.}

\item{y}{bl or bl_data object to adjust by.}
}
\value{
Adjusted bl object.
}
\description{
Bayes linear kinematic adjustment
}
