% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_resolution.R
\name{bl_resolution}
\alias{bl_resolution}
\title{Bayes linear kinematic adjustment}
\usage{
bl_resolution(x, y)
}
\arguments{
\item{x}{bl object prior to adjustment.}

\item{y}{bl object after adjustment.}
}
\value{
Vector of resolutions on the 0,1 scale.
}
\description{
Bayes linear kinematic adjustment
}
\examples{
bl1 <- bl(name = 'Example 1',
         varnames = c('x', 'y', 'z'),
         expectation = c(1, 2, 3),
         covariance = matrix(c(1, 0.5, 0.5,
           0.5, 1, 0.5,
           0.5, 0.5, 1), 3, 3))
           
bl2 <- bl(name = 'Example 2',
        varnames = c('x', 'y', 'z'),
        expectation = c(1.1, 2.1, 3.1),
        covariance = matrix(c(1.1, 0.6, 0.6,
        0.6, 1.1, 0.6,
        0.6, 0.6, 1.1), 3, 3))
        
bl_resolution(bl2, bl1)         
}
